fix: 모달 기능 분리 및 recoil 사용
