+++ package.json
@@ -18,6 +18,7 @@
     "react-modal": "^3.16.1",
     "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
+    "recoil": "^0.7.7",
     "styled-components": "^5.3.9",
     "styled-normalize": "^8.0.7",
     "typescript": "^4.9.5",
